requirements.txt
pygetwindow
pillow
numpy 
opencv-python 
keyboard 
pyautogui